Decoy generation for compose JS runtime
Based on the exploration and the fixes from KT-40980, the best idea to fix the problem with transformed @Composable functions was to generate duplicated functions which will be the target for all transforms.

This CL does exactly this, hiding decoys behind a compilation flag for now.

To run, use `./gradlew :compose:runtime:runtime:clean :compose:runtime:runtime:jsJar -Pandroidx.compose.multiplatformEnabled=true`

Change-Id: Ie0485aab4b69717e679acafc5e6d83dacf2a86c6